Adding R Project that Computes Metrics
R project and associated files are in rmetrics folder. You will need to update the file path in the metrics script to import the csv into your R environment. brandvideof.csv is the frame-level metrics for the NFL game. brandmetrics.csv is the video-level metrics for every brand in the NFL game.